+++ CHANGELOG.md
@@ -204,6 +204,7 @@ as necessary. Empty sections will not end in the release notes.
 ### New Features
 
 - CLI: New `REVERT CONTENT` command to update one or more tables or views to a previous state.
+- Catalog: Add secrets providers for AWS and Vault, experimental for Azure + GCP
 
 ### Changes